+++ src/components/Footer.tsx
@@ -1,5 +1,11 @@
 import Image from 'next/image';
 import Link from 'next/link';
+import LogoName from '@/../public/logo-name-white.svg';
+import Arrow from '@/../public/icons/arrow-white.svg';
+import Discord from '@/../public/icons/join-discord-white.svg';
+import Instagram from '@/../public/icons/instagram-white.svg';
+import LinkedIn from '@/../public/icons/linkedin-white.svg';
+import GitHub from '@/../public/icons/github-white.svg';
 
 const linkClasses = 'underline decoration-transparent hover:decoration-inherit transition';
 
@@ -7,18 +13,12 @@ const Footer = () => {
   return (
     <footer className="bg-royal text-white w-full rounded-t-[3.125rem] pb-10 md:px-40 px-8 text-sm mt-20">
       <div className="flex justify-between py-16 sm:py-28 gap-8">
-        <Image
-          src="/logo-name-white.svg"
-          alt="logo"
-          width="360"
-          height="53"
-          className="shrink min-w-0"
-        />
+        <Image src={LogoName} alt="logo" width="360" height="53" className="shrink min-w-0" />
         <button
           onClick={() => window.scrollTo(0, 0)}
           className="items-center flex flex-col rounded-full p-2 transition border border-white/0 hover:border-white"
         >
-          <Image src="/arrow-white.svg" alt="arrow" width="20" height="20" className="rotate-180" />
+          <Image src={Arrow} alt="arrow" width="20" height="20" className="rotate-180" />
           Top
         </button>
       </div>
@@ -81,14 +81,14 @@ const Footer = () => {
             href="https://discord.gg/tcpcnfxmeQ"
             target="_blank"
           >
-            <Image src="/join-discord-white.svg" alt="discord" width="200" height="60" />
+            <Image src={Discord} alt="discord" width="200" height="60" />
           </a>
           <a
             className={linkClasses + ' flex items-center gap-2'}
             href="https://instagram.com/utdnebula"
             target="_blank"
           >
-            <Image src="/instagram-white.svg" alt="Instagram logo" width="30" height="30" />
+            <Image src={Instagram} alt="Instagram logo" width="30" height="30" />
             Instagram
           </a>
           <a
@@ -96,7 +96,7 @@ const Footer = () => {
             href="https://linkedin.com/company/utdnebula"
             target="_blank"
           >
-            <Image src="/linkedin-white.svg" alt="LinkedIn logo" width="30" height="30" />
+            <Image src={LinkedIn} alt="LinkedIn logo" width="30" height="30" />
             LinkedIn
           </a>
           <a
@@ -104,7 +104,7 @@ const Footer = () => {
             href="https://github.com/utdnebula"
             target="_blank"
           >
-            <Image src="/github-white.svg" alt="GitHub logo" width="30" height="30" />
+            <Image src={GitHub} alt="GitHub logo" width="30" height="30" />
             GitHub
           </a>
         </div>
