+++ src/components/Navbar.tsx
@@ -4,18 +4,18 @@ import type { StaticImageData } from 'next/image';
 import Link from 'next/link';
 import { Disclosure, Transition, TransitionRootProps } from '@headlessui/react';
 import { Fragment, useCallback, useEffect, useRef, useState } from 'react';
-import FilledChevronUpWhite from '@/../public/filled-chevron-up-white.svg';
-import FilledChevronUpRoyal from '@/../public/filled-chevron-up-royal.svg';
-import Arrow from '@/../public/arrow-white.svg';
-import X from '@/../public/x.svg';
-import HamburgerWhite from '@/../public/menu-white.svg';
-import HamburgerRoyal from '@/../public/menu-royal.svg';
-import Puzzle from '@/../public/puzzle.svg';
-import Users from '@/../public/users.svg';
-import UserGroup from '@/../public/user-group.svg';
-import TrendingUp from '@/../public/trending-up.svg';
-import Pencil from '@/../public/pencil.svg';
-import Star from '@/../public/star.svg';
+import FilledChevronUpWhite from '@/../public/icons/filled-chevron-up-white.svg';
+import FilledChevronUpRoyal from '@/../public/icons/filled-chevron-up-royal.svg';
+import Arrow from '@/../public/icons/arrow-white.svg';
+import X from '@/../public/icons/x.svg';
+import HamburgerWhite from '@/../public/icons/menu-white.svg';
+import HamburgerRoyal from '@/../public/icons/menu-royal.svg';
+import Puzzle from '@/../public/icons/puzzle.svg';
+import Users from '@/../public/icons/users.svg';
+import UserGroup from '@/../public/icons/user-group.svg';
+import TrendingUp from '@/../public/icons/trending-up.svg';
+import Pencil from '@/../public/icons/pencil.svg';
+import Star from '@/../public/icons/star.svg';
 
 type BaseItem = { name: string };
