+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import Navbar from '@/components/Navbar';
 import Image from 'next/image';
-import Circles from '@/../public/circles.svg';
+import Circles from '@/../public/images/circles.svg';
 import Blob from '@/components/Blob';
 import Footer from '@/components/Footer';
-import Arrow from '@/../public/arrow-white.svg';
+import Arrow from '@/../public/icons/arrow-white.svg';
 import WhoWeAre from '@/components/WhoWeAre';
 import useBlobBg from '@/lib/useBlobBg';
 import CTA from '@/components/CTA';
